requirements.txt
fastapi
spacy
pydantic
pdfminer
requestspip